+++ apps/web/src/app/tournament/join/[id]/_components/JoinErrorScreen.tsx
@@ -1,8 +1,8 @@
-import TournamentErrorDialog from '@/components/tournament-error-dialog';
-import type { TournamentErrorTypeT } from '@/types/tournament';
+import JoinErrorDialog from '@/components/common/join-error-dialog';
+import type { JoinErrorTypeT } from '@/components/common/join-error-dialog';
 
 type JoinErrorScreenProps = {
-  type: TournamentErrorTypeT;
+  type: JoinErrorTypeT;
 };
 
 /** 토너먼트 참여 불가 안내 화면 */
@@ -10,7 +10,7 @@ function JoinErrorScreen({ type }: JoinErrorScreenProps) {
   return (
     <>
       <main className="min-h-dvh bg-bg-layer-default" />
-      <TournamentErrorDialog type={type} />
+      <JoinErrorDialog type={type} />
     </>
   );
 }
